requirements.txt
streamlit
RPi.GPIO
adafruit-circuitpython-dht
blinka
